requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
openai>=1.0.0
tenacity>=8.2.0
aiohttp>=3.8.5
redis>=4.5.1
hiredis>=2.2.3
uuid==1.30
